Add metrics Label generation to MessageSourceGenerator
The Label property is what we will use to label metrics about Messages
The LabelStatic property is so that we can set the label for the whole class of Messages. (via configuration in a future PR)
The OriginalLabelStatic property is so that we can always see the original label before the configuration changes it.